Fixed bugs related to database inconsistencies between 32 bit and 64 bit.

+++ engine/server/library/serverGame/src/shared/object/ResourcePoolObject.cpp
@@ -54,7 +54,7 @@ ResourcePoolObject::~ResourcePoolObject()
  */
 float ResourcePoolObject::harvest(float installedEfficiency, uint32 lastHarvestTime) const
 {
-	int elapsedTime=static_cast<int>(std::min(ServerClock::getInstance().getGameTimeSeconds(),static_cast<unsigned long>(m_depletedTimestamp)) - lastHarvestTime);
+	uint32_t elapsedTime=static_cast<uint32_t>(std::min(ServerClock::getInstance().getGameTimeSeconds(),static_cast<uint32_t>(m_depletedTimestamp)) - lastHarvestTime);
 	if (elapsedTime < 0)
 		return 0;
 	float elapsedTicks =static_cast<float>(elapsedTime) / static_cast<float>(ConfigServerGame::getSecondsPerResourceTick());
